move upload handler and related functions to new type...
also
* update uploads config directory path to be pathbuf
