+++ src/http_api/mod.rs
@@ -1,4 +1,8 @@
-pub mod data;
+mod data;
 mod entry;
+mod errors;
 mod mock;
+mod upload;
+
 pub use entry::*;
+pub use upload::UploadConfig;
